Move permission test to separete class
